Multiple docstrings improvements
