+++ src/ansys/aedt/core/modeler/cad/object_3d.py
@@ -379,7 +379,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
 
         References
         ----------
@@ -405,7 +405,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
         """
         f_list = []
         for face in self.faces:
@@ -448,7 +448,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
         """
         f = []
         for face in self.faces:
@@ -463,7 +463,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
         """
         e = []
         for edge in self.edges:
@@ -478,7 +478,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
         """
         f = []
         for face in self.faces:
@@ -493,7 +493,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
         """
         e = []
         for edge in self.edges:
@@ -737,7 +737,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
 
         References
         ----------
@@ -759,7 +759,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.VertexPrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.VertexPrimitive`]
 
         References
         ----------
@@ -1609,7 +1609,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.object_3d.Object3d`]
+        list[:class:`ansys.aedt.core.modeler.cad.object_3d.Object3d`]
             List of names of the newly added objects.
 
         References
@@ -1779,7 +1779,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.object_3d.Object3d`]
+        list[:class:`ansys.aedt.core.modeler.cad.object_3d.Object3d`]
             List of object resulting from the operation.
 
         References
@@ -1895,7 +1895,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.FacePrimitive`]
             List of face primitives.
         """
 
@@ -1939,7 +1939,7 @@ class Object3d(object):
 
         Returns
         -------
-        List[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
+        list[:class:`ansys.aedt.core.modeler.cad.elements_3d.EdgePrimitive`]
             List of edge primitives.
         """
         filters = ["==", "<=", ">=", "<", ">"]
